requirements.txt
allure-pytest
allure-python-commons
pytest-rerunfailure
altgraph
atomicwrites
attrs
brotlipy
certifi
colorama
cx-Oracle
Faker
future
greenlet
idna
iniconfig
jdcal
kafka
kafka-python
lxml
openpyxl
packaging
pefile
pluggy
psycopg2
pycosat
pycparser
pyinstaller
pyinstaller-hooks-contrib
pytest
pytest-dependency
pytest-json-report
pytest-metadata
pytest-mock
pytest-ordering
pytest-rerunfailures
python-dateutil
PyYAML
records
requests
pytest-rerunfailure